requirements.txt
fastapi == 0.115.5
pydantic == 2.10.1
uvicorn == 0.32.1
python-dotenv == 1.0.1
textblob == 0.19.0
streamlit == 1.41.1
sqlite == 3.45.3